requirements.txt
Django==3.2
python-dotenv==0.18.0
psycopg2==2.9.5
celery==5.2.7
redis==4.4.0
django-celery-beat==2.3.0
flower==1.2.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
Markdown==3.4.1
django-filter==22.1
pika==1.3.1
